Work In Progress
I'm getting closer to properly parsing the HTML returned.

For some reason curl returns the processed HTML but HTTPoison doesn't seem to.

I've gotten to the point that I can parse out the date of the event (day, month and year) but I still need to figure out how to combine that information with the time and, of course, the name of the event.